+++ src/main/java/com/diploma/edu/source/controllers/BuildingController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -28,19 +29,6 @@ public class BuildingController {
 
     @GetMapping
     public Page<Building> getAll(@RequestParam Map<String, String> params) {
-
-//        List<SearchCriteria> filters = new ArrayList<>();
-//        Pageable pageable = null;
-//        if (page == null && size != null) {
-//            pageable = PageRequest.of(0, size);
-//        }
-//        if (page != null && size != null) {
-//            pageable = PageRequest.of(page, size);
-//        }
-//        if (number != null) {
-//            filters.add(new SearchCriteria("number", number));
-//        }
-//        return service.getAll(pageable, filters, new SortCriteria(sort));
         return service.getAll(GetRequestParams.getPageable(params),
                 GetRequestParams.getFilters(params),
                 GetRequestParams.getSortCriteria(params));
@@ -52,12 +40,12 @@ public class BuildingController {
     }
 
     @GetMapping("{id}")
-    public Building getBuildingById(@PathVariable("id") Long buildingID) {
+    public Building getBuildingById(@PathVariable("id") BigInteger buildingID) {
         return service.getById(buildingID);
     }
 
     @DeleteMapping("{id}")
-    public boolean deleteBuilding(@PathVariable("id") Long buildingID) {
+    public boolean deleteBuilding(@PathVariable("id") BigInteger buildingID) {
         return service.delete(buildingID);
     }
