+++ src/main/java/com/diploma/edu/source/controllers/ElectronicKeyController.java
@@ -7,6 +7,7 @@ import com.diploma.edu.source.servicies.requestBuilder.criteria.SortCriteria;
 import org.springframework.data.domain.*;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -23,29 +24,6 @@ public class ElectronicKeyController {
 
     @GetMapping
     public Page<Ekey> getAll(@RequestParam Map<String, String> params) {
-
-//        List<SearchCriteria> filters = new ArrayList<>();
-//        Pageable pageable = null;
-//        if (page == null && size != null) {
-//            pageable = PageRequest.of(0, size);
-//        }
-//        if (page != null && size != null) {
-//            pageable = PageRequest.of(page, size);
-//        }
-//        if (keyCode != null) {
-//            filters.add(new SearchCriteria("keyCode", "like '%" + keyCode + "%' "));
-//        }
-//        if (name != null) {
-//            filters.add(new SearchCriteria("name", "like '%" + name + "%' "));
-//        }
-//        if (isActive != null) {
-//            filters.add(new SearchCriteria("isActive", "like '%" + isActive + "%' "));
-//        }
-//        if (userId != null) {
-//            filters.add(new SearchCriteria("user", userId));
-//        }
-//        return service.getAll(pageable, filters, new SortCriteria(sort));
-
         return service.getAll(GetRequestParams.getPageable(params),
                 GetRequestParams.getFilters(params),
                 GetRequestParams.getSortCriteria(params));
@@ -57,7 +35,7 @@ public class ElectronicKeyController {
     }
 
     @DeleteMapping("{id}")
-    public boolean deleteKey(@PathVariable("id") Long keyID) {
+    public boolean deleteKey(@PathVariable("id") BigInteger keyID) {
         return service.delete(keyID);
     }
 
@@ -67,7 +45,7 @@ public class ElectronicKeyController {
     }
 
     @RequestMapping(value = "/get-one/{id}")
-    public Ekey getOne(@PathVariable("id") Long id) {
+    public Ekey getOne(@PathVariable("id") BigInteger id) {
         return service.getById(id);
     }
