+++ src/main/java/com/diploma/edu/source/controllers/LoggerController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -24,32 +25,13 @@ public class LoggerController {
 
     @GetMapping
     public Page<Logger> getAll(@RequestParam Map<String, String> params) {
-
-//        List<SearchCriteria> filters = new ArrayList<>();
-//        Pageable pageable = null;
-//        if (page == null && size != null) {
-//            pageable = PageRequest.of(0, size);
-//        }
-//        if (page != null && size != null) {
-//            pageable = PageRequest.of(page, size);
-//        }
-//        if (dateFrom != null) {
-//            filters.add(new SearchCriteria("dateAndTime", " > to_date('" + changeDateFormat(new Date(dateFrom))
-//                    + "', 'yyyy-mm-dd hh24:mi:ss')"));
-//        }
-//        if (dateTo != null) {
-//            filters.add(new SearchCriteria("dateAndTime", " < to_date('" + changeDateFormat(new Date(dateTo))
-//                    + "', 'yyyy-mm-dd hh24:mi:ss')"));
-//        }
-//        return loggerService.getAll(pageable, filters, new SortCriteria(sort));
-
         return service.getAll(GetRequestParams.getPageable(params),
                 GetRequestParams.getFilters(params),
                 GetRequestParams.getSortCriteria(params));
     }
 
     @GetMapping("{id}")
-    public Logger getLogger(@PathVariable("id") Long loggerID) {
+    public Logger getLogger(@PathVariable("id") BigInteger loggerID) {
         return service.getById(loggerID);
     }
