+++ src/main/java/com/diploma/edu/source/controllers/RoleController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -28,20 +29,6 @@ public class RoleController {
 
     @GetMapping
     public Page<Role> getAll(@RequestParam Map<String, String> params) {
-
-//        List<SearchCriteria> filters = new ArrayList<>();
-//        Pageable pageable = null;
-//        if (page == null && size != null) {
-//            pageable = PageRequest.of(0, size);
-//        }
-//        if (page != null && size != null) {
-//            pageable = PageRequest.of(page, size);
-//        }
-//        if (role != null) {
-//            filters.add(new SearchCriteria("flat", role));
-//        }
-//        return service.getAll(pageable, filters, new SortCriteria(sort));
-
         return service.getAll(GetRequestParams.getPageable(params),
                 GetRequestParams.getFilters(params),
                 GetRequestParams.getSortCriteria(params));
@@ -53,7 +40,7 @@ public class RoleController {
     }
 
     @DeleteMapping("{id}")
-    public boolean deleteRole(@PathVariable("id") Long roleId) {
+    public boolean deleteRole(@PathVariable("id") BigInteger roleId) {
         return service.delete(roleId);
     }
