+++ src/main/java/com/diploma/edu/source/controllers/UtilitiesController.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.*;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.*;
@@ -31,56 +32,13 @@ public class UtilitiesController {
 
     @GetMapping
     public Page<Utility> getAll(@RequestParam Map<String, String> params) {
-
-//        List<SearchCriteria> filters = new ArrayList<>();
-//        Pageable pageable = null;
-//        if (page == null && size != null) {
-//            pageable = PageRequest.of(0, size);
-//        }
-//        if (page != null && size != null) {
-//            pageable = PageRequest.of(page, size);
-//        }
-//        if (bankBook != null) {
-//            filters.add(new SearchCriteria("bankBook", "like '%" + bankBook + "%' "));
-//        }
-//        if (dateFrom != null) {
-//            filters.add(new SearchCriteria("dateAndTime", " > to_date('" + changeDateFormat(new Date(dateFrom))
-//                    + "', 'yyyy-mm-dd hh24:mi:ss')"));
-//        }
-//        if (dateTo != null) {
-//            filters.add(new SearchCriteria("dateAndTime", " < to_date('" + changeDateFormat(new Date(dateTo))
-//                    + "', 'yyyy-mm-dd hh24:mi:ss')"));
-//        }
-//        if (date != null) {
-//            filters.add(new SearchCriteria("month", getMonthAndYear(new Date(date))));
-//        }
-//        if (name != null) {
-//            filters.add(new SearchCriteria("name", "like '%" + name + "%' "));
-//        }
-//        if (currentMonthReading != null) {
-//            filters.add(new SearchCriteria("currentMonthReading", "like '%" + currentMonthReading + "%' "));
-//        }
-//        if (currentMonthReading != null) {
-//            filters.add(new SearchCriteria("lastMonthReading", "like '%" + lastMonthReading + "%' "));
-//        }
-//        if (status != null) {
-//            filters.add(new SearchCriteria("status", "like '%" + status + "%' "));
-//        }
-//        if (serviceID != null) {
-//            filters.add(new SearchCriteria("service", serviceID));
-//        }
-//        if (address != null) {
-//            filters.add(new SearchCriteria("address", address));
-//        }
-//        return service.getAll(pageable, filters, new SortCriteria(sort));
-
         return service.getAll(GetRequestParams.getPageable(params),
                 GetRequestParams.getFilters(params),
                 GetRequestParams.getSortCriteria(params));
     }
 
     @GetMapping("{id}")
-    public Utility getUtility(@PathVariable("id") Long utilityID) {
+    public Utility getUtility(@PathVariable("id") BigInteger utilityID) {
         return service.getById(utilityID);
     }
 
@@ -96,12 +54,13 @@ public class UtilitiesController {
 
     @PutMapping
     public boolean updateUtility(@RequestBody Utility utility) {
+        //TODO: заменить проверку и расчет
         if (utility.getEndMonthReading() == null || utility.getStartMonthReading() == null) {
             Utility utilityFromDB = service.getById(utility.getId());
             if (utility.getEndMonthReading() != null) {
-                if (utility.getEndMonthReading() > utilityFromDB.getStartMonthReading()) {
+                if (utility.getEndMonthReading().intValue() > utilityFromDB.getStartMonthReading().intValue()) {
                     utility.setAmountToPay(
-                            (utility.getEndMonthReading() - utility.getStartMonthReading())
+                            (utility.getEndMonthReading().intValue() - utility.getStartMonthReading().intValue())
                                     * utility.getService().getTariff()
                     );
                     utility.setStatus(true);
