+++ src/main/java/com/diploma/edu/source/model/BaseEntity.java
@@ -1,17 +1,18 @@
 package com.diploma.edu.source.model;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 public abstract class BaseEntity implements Serializable {
-    protected Long id;
+    protected BigInteger id;
     protected String name;
     protected String description;
 
-    public Long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
