+++ src/main/java/com/diploma/edu/source/servicies/EkeyService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @org.springframework.stereotype.Service
@@ -21,7 +22,7 @@ public class EkeyService implements Service<Ekey> {
     }
 
     @Override
-    public Ekey getById(Long id) {
+    public Ekey getById(BigInteger id) {
         return oracleDbAccess.getById(Ekey.class, id);
     }
 
@@ -35,7 +36,7 @@ public class EkeyService implements Service<Ekey> {
     }
 
     @Override
-    public boolean delete(Long id) {
+    public boolean delete(BigInteger id) {
         if (oracleDbAccess.delete(Ekey.class, id) == 1) {
             return false;
         } else {
