+++ src/main/java/com/diploma/edu/source/servicies/Service.java
@@ -5,15 +5,17 @@ import com.diploma.edu.source.servicies.requestBuilder.criteria.SortCriteria;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface Service<T> {
 
-    T getById(Long id);
+    T getById(BigInteger id);
 
     boolean create(T object);
 
-    boolean delete(Long id);
+    boolean delete(BigInteger id);
 
     boolean update(T object);
